backend/requirements.txt
Flask
pyairtable
flask-cors
mercadopago
python-dotenv
gunicorn
resend
weasyprint
resend
